requirements.txt
alabaster==0.7.12
asgiref==3.2.10
attrs==20.2.0
Babel==2.8.0
bcrypt==3.2.0
cachetools==4.1.1
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
chembl-webresource-client==0.10.2
cryptography==3.1
decorator==4.4.2
defusedxml==0.7.0rc1
Django==3.1.1
django-crispy==1.8.1
djangorestframework==3.11.1
docutils==0.16
easydict==1.9
ecdsa==0.16.0
httplib2==0.18.1
idna==2.10
imagesize==1.2.0
Jinja2==2.11.2
jsonpath-ng==1.5.2
jsonschema==3.2.0
MarkupSafe==1.1.1
mypy-extensions==0.4.3
mysqlclient==1.4.6
oauth2client==4.1.3
oauthlib==3.1.0
packaging==20.4
paramiko==2.7.2
ply==3.11
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pycurl==7.43.0.6
Pygments==2.7.1
PyJWT==1.7.1
PyNaCl==1.4.0
pyparsing==2.4.7
pyrsistent==0.17.3
pysftp==0.2.9
python-dotenv==0.14.0
python-jose==3.2.0
python3-openid==3.2.0
pytz==2020.1
PyYAML==5.3.1
qwikidata==0.4.0
requests==2.24.0
requests-cache==0.5.2
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
slacker==0.14.0
snowballstemmer==2.0.0
social-auth-app-django==4.0.0
social-auth-core==3.3.3
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
uritemplate==3.0.1
urllib3==1.25.10
